+++ metadata-io/src/main/java/com/linkedin/metadata/test/eval/operator/LessThanEvaluator.java
@@ -65,7 +65,7 @@ public class LessThanEvaluator extends BaseOperatorEvaluator {
     try {
       Number lhsNumber = NumberFormat.getInstance().parse(lhs);
       Number rhsNumber = NumberFormat.getInstance().parse(rhs);
-      return lhsNumber.intValue() < rhsNumber.intValue(); // REVISIT THIS!
+      return lhsNumber.longValue() < rhsNumber.longValue();
     } catch (ParseException e) {
       log.warn(
           "Failed to evaluate Less Than (<) Operator. Input values "
